Initial work from static classes to real objects
This commit is the beginning of moving all static classes into real
objects. It also includes work on renaming "plugins" to "helpers"
which is used within controllers for various things.

This commit has focused on core functionality and some helpers will
probably not work before properly implemented.

This commit also introduces the dispatcher class which does the execution
steps that used to be in the hook object.