Theme object changes
The theme class is not bound to the controller and allows the controller
to add CSS/JS to it dynamically. Default CSS/JS for the theme is added
to its theme.json file.

Theme class is no longer static and now adds variables to templates that is
set on the object. Currently only CSS/JS is added to this container.